Updated GamePanel.java for Score & Directions
Changed the paintComponent method so that a score is constantly shown in the game in the top left hand corner of the screen.
This score is incremented in the checkForCollisions method in Game.java.

Additionally, I added directions to the player at the beginning of the game before the player starts (tells the player what buttons can be pressed) and when the game ends (tells the player how to restart the game).

Before this commit, I had no idea (without looking in the code or pressing random keys) how to start the game. So, I figured these directions would be useful to the player.